requirements.txt
aniso8601 (3.0.2)
appnope (0.1.0)
backcall (0.1.0)
certifi (2018.4.16)
chardet (3.0.4)
coreapi (2.3.3)
coreschema (0.0.4)
decorator (4.3.0)
diff-match-patch (20121119)
Django (2.0.7)
django-coreapi (2.3.0)
django-cors-headers (2.4.0)
django-crispy-forms (1.7.2)
django-environ (0.4.5)
django-filter (1.1.0)
django-formtools (2.1)
django-graphql-jwt (0.1.10)
django-guardian (1.4.9)
django-import-export (1.0.1)
django-reversion (3.0.0)
djangorestframework (3.8.2)
djangorestframework-jwt (1.11.0)
et-xmlfile (1.0.1)
future (0.16.0)
graphene (2.1.3)
graphene-django (2.1.0)
graphql-core (2.1)
graphql-relay (0.4.5)
httplib2 (0.9.2)
idna (2.7)
ipython (6.4.0)
ipython-genutils (0.2.0)
iso8601 (0.1.12)
itypes (1.1.0)
jdcal (1.4)
jedi (0.12.1)
Jinja2 (2.10)
Markdown (2.6.11)
MarkupSafe (1.0)
mysqlclient (1.3.13)
odfpy (1.3.6)
openpyxl (2.5.4)
parso (0.3.1)
pexpect (4.6.0)
pickleshare (0.7.4)
Pillow (5.2.0)
pip (9.0.1)
promise (2.1)
prompt-toolkit (1.0.15)
ptyprocess (0.6.0)
Pygments (2.2.0)
PyJWT (1.6.4)
pytz (2018.5)
PyYAML (3.13)
requests (2.19.1)
Rx (1.6.1)
setuptools (28.8.0)
simplegeneric (0.8.1)
singledispatch (3.4.0.3)
six (1.11.0)
tablib (0.12.1)
traitlets (4.3.2)
typing (3.6.4)
unicodecsv (0.14.1)
uritemplate (3.0.0)
urllib3 (1.23)
wcwidth (0.1.7)
xadmin (2.0.1)
xlrd (1.1.0)
xlwt (1.3.0)
